Moved DoneButton to its own class.
