+++ src/lib/Components/Consignments/Screens/Artist.tsx
@@ -1,8 +1,8 @@
 import * as React from "react"
 
 import ArtistSearch from "../Components/ArtistSearchResults"
-import DoneButton from "../Components/BottomAlignedButton"
 import ConsignmentBG from "../Components/ConsignmentBG"
+import DoneButton from "../Components/DoneButton"
 
 import { ArtistResult, ConsignmentSetup } from "../index"
 
@@ -66,17 +66,9 @@ export default class Artist extends React.Component<Props, State> {
   }
 
   render() {
-    // This might become a higher order component for reuse, if used more elsewhere
-    const doneButtonStyles = {
-      backgroundColor: "black",
-      marginBottom: 20,
-      paddingTop: 18,
-      height: 56,
-    }
-
     return (
       <ConsignmentBG>
-        <DoneButton onPress={this.doneTapped} bodyStyle={doneButtonStyles} buttonText="DONE">
+        <DoneButton onPress={this.doneTapped}>
           <View
             style={{ alignContent: "center", justifyContent: "flex-end", flexGrow: 1, marginLeft: 20, marginRight: 20 }}
           >
